[#11] - Move `Header` and `Main` to separate folder
